+++ Sources/ParseSwift/Types/ParsePushStatus.swift
@@ -15,9 +15,9 @@ import Foundation
  use the master key in server-side applications where the key is kept secure and not
  exposed to the public.
  */
-public struct ParsePushStatus<U: ParseInstallation, V: ParsePushPayloadDatable>: ParsePushStatusable {
+public struct ParsePushStatus<U: ParseInstallation, V: ParsePushPayloadable>: ParsePushStatusable {
     public typealias InstallationType = U
-    public typealias PushType = V
+    public typealias PayloadType = V
 
     public var originalData: Data?
 
@@ -35,7 +35,7 @@ public struct ParsePushStatus<U: ParseInstallation, V: ParsePushPayloadDatable>:
 
     public var source: String?
 
-    public var payload: ParsePushPayload<PushType>?
+    public var payload: PayloadType?
 
     public var title: String?
